+++ cli/internal/i18n/i18n.go
@@ -35,4 +35,3 @@ func T(msg Message) string {
 	}
 	return msg.EN
 }
-